fixed floating negotiation values issue that prevented the users from submitting the negotiation if the value contains floating values

+++ lib/features/home/widgets/send_offer_screen.dart
@@ -499,7 +499,7 @@ class _SendOfferScreenState extends State<SendOfferScreen> {
                                       selectedBorder.hasValue &&
                                       selectedBranch.hasValue) {
                                     negotiationBloc.sendOffer(SendOffer(
-                                      price: int.parse(controller.text),
+                                      price: double.parse(controller.text),
                                       requestId: widget.result.id,
                                       providerId: user.id,
                                       periodId: selectedBorder.value,
